Start breaking ChatListRenderer content types into their own classes
