+++ src/vs/workbench/contrib/chat/browser/chatContentParts/chatCollections.ts
@@ -1,0 +1,43 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { IDisposable, Disposable } from 'vs/base/common/lifecycle';
+
+export class ResourcePool<T extends IDisposable> extends Disposable {
+	private readonly pool: T[] = [];
+
+	private _inUse = new Set<T>;
+	public get inUse(): ReadonlySet<T> {
+		return this._inUse;
+	}
+
+	constructor(
+		private readonly _itemFactory: () => T,
+	) {
+		super();
+	}
+
+	get(): T {
+		if (this.pool.length > 0) {
+			const item = this.pool.pop()!;
+			this._inUse.add(item);
+			return item;
+		}
+
+		const item = this._register(this._itemFactory());
+		this._inUse.add(item);
+		return item;
+	}
+
+	release(item: T): void {
+		this._inUse.delete(item);
+		this.pool.push(item);
+	}
+}
+
+export interface IDisposableReference<T> extends IDisposable {
+	object: T;
+	isStale: () => boolean;
+}
