+++ src/vs/workbench/contrib/chat/browser/chatContentParts/chatConfirmationWidget.ts
