+++ src/vs/workbench/contrib/chat/browser/chatContentParts/chatTreeContentPart.ts
@@ -1,0 +1,223 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import * as dom from 'vs/base/browser/dom';
+import { IListVirtualDelegate } from 'vs/base/browser/ui/list/list';
+import { ITreeCompressionDelegate } from 'vs/base/browser/ui/tree/asyncDataTree';
+import { ICompressedTreeNode } from 'vs/base/browser/ui/tree/compressedObjectTreeModel';
+import { ICompressibleTreeRenderer } from 'vs/base/browser/ui/tree/objectTree';
+import { IAsyncDataSource, ITreeNode } from 'vs/base/browser/ui/tree/tree';
+import { Emitter, Event } from 'vs/base/common/event';
+import { Disposable, DisposableStore } from 'vs/base/common/lifecycle';
+import { localize } from 'vs/nls';
+import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
+import { FileKind, FileType } from 'vs/platform/files/common/files';
+import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
+import { WorkbenchCompressibleAsyncDataTree } from 'vs/platform/list/browser/listService';
+import { IOpenerService } from 'vs/platform/opener/common/opener';
+import { IThemeService } from 'vs/platform/theme/common/themeService';
+import { IResourceLabel, ResourceLabels } from 'vs/workbench/browser/labels';
+import { ChatTreeItem } from 'vs/workbench/contrib/chat/browser/chat';
+import { IDisposableReference, ResourcePool } from 'vs/workbench/contrib/chat/browser/chatContentParts/chatCollections';
+import { IChatCommandButton, IChatResponseProgressFileTreeData } from 'vs/workbench/contrib/chat/common/chatService';
+import { createFileIconThemableTreeContainerScope } from 'vs/workbench/contrib/files/browser/views/explorerView';
+import { IFilesConfiguration } from 'vs/workbench/contrib/files/common/files';
+
+const $ = dom.$;
+
+export class ChatTreeContentPart extends Disposable {
+	public readonly element: HTMLElement;
+
+	private readonly _onDidChangeHeight = this._register(new Emitter<void>());
+	public readonly onDidChangeHeight = this._onDidChangeHeight.event;
+
+	public readonly onDidFocus: Event<void>;
+
+	private tree: WorkbenchCompressibleAsyncDataTree<IChatResponseProgressFileTreeData, IChatResponseProgressFileTreeData, void>;
+
+	constructor(
+		data: IChatResponseProgressFileTreeData,
+		element: ChatTreeItem,
+		treePool: TreePool,
+		treeDataIndex: number,
+		@IOpenerService private readonly openerService: IOpenerService
+	) {
+		super();
+
+		const ref = this._register(treePool.get());
+		this.tree = ref.object;
+		this.onDidFocus = this.tree.onDidFocus;
+
+		this._register(this.tree.onDidOpen((e) => {
+			if (e.element && !('children' in e.element)) {
+				this.openerService.open(e.element.uri);
+			}
+		}));
+		this._register(this.tree.onDidChangeCollapseState(() => {
+			this._onDidChangeHeight.fire();
+		}));
+		this._register(this.tree.onContextMenu((e) => {
+			e.browserEvent.preventDefault();
+			e.browserEvent.stopPropagation();
+		}));
+
+		this.tree.setInput(data).then(() => {
+			if (!ref.isStale()) {
+				this.tree.layout();
+				this._onDidChangeHeight.fire();
+			}
+		});
+
+		this.element = this.tree.getHTMLElement().parentElement!;
+	}
+
+	domFocus() {
+		this.tree.domFocus();
+	}
+
+	isDifferent(command: IChatCommandButton): boolean {
+		// This chat type can't be updated in the model
+		return false;
+	}
+
+	update(command: IChatCommandButton): void {
+		// This chat type can't be updated in the model
+	}
+}
+
+export class TreePool extends Disposable {
+	private _pool: ResourcePool<WorkbenchCompressibleAsyncDataTree<IChatResponseProgressFileTreeData, IChatResponseProgressFileTreeData, void>>;
+
+	public get inUse(): ReadonlySet<WorkbenchCompressibleAsyncDataTree<IChatResponseProgressFileTreeData, IChatResponseProgressFileTreeData, void>> {
+		return this._pool.inUse;
+	}
+
+	constructor(
+		private _onDidChangeVisibility: Event<boolean>,
+		@IInstantiationService private readonly instantiationService: IInstantiationService,
+		@IConfigurationService private readonly configService: IConfigurationService,
+		@IThemeService private readonly themeService: IThemeService,
+	) {
+		super();
+		this._pool = this._register(new ResourcePool(() => this.treeFactory()));
+	}
+
+	private treeFactory(): WorkbenchCompressibleAsyncDataTree<IChatResponseProgressFileTreeData, IChatResponseProgressFileTreeData, void> {
+		const resourceLabels = this._register(this.instantiationService.createInstance(ResourceLabels, { onDidChangeVisibility: this._onDidChangeVisibility }));
+
+		const container = $('.interactive-response-progress-tree');
+		this._register(createFileIconThemableTreeContainerScope(container, this.themeService));
+
+		const tree = this.instantiationService.createInstance(
+			WorkbenchCompressibleAsyncDataTree<IChatResponseProgressFileTreeData, IChatResponseProgressFileTreeData>,
+			'ChatListRenderer',
+			container,
+			new ChatListTreeDelegate(),
+			new ChatListTreeCompressionDelegate(),
+			[new ChatListTreeRenderer(resourceLabels, this.configService.getValue('explorer.decorations'))],
+			new ChatListTreeDataSource(),
+			{
+				collapseByDefault: () => false,
+				expandOnlyOnTwistieClick: () => false,
+				identityProvider: {
+					getId: (e: IChatResponseProgressFileTreeData) => e.uri.toString()
+				},
+				accessibilityProvider: {
+					getAriaLabel: (element: IChatResponseProgressFileTreeData) => element.label,
+					getWidgetAriaLabel: () => localize('treeAriaLabel', "File Tree")
+				},
+				alwaysConsumeMouseWheel: false
+			});
+
+		return tree;
+	}
+
+	get(): IDisposableReference<WorkbenchCompressibleAsyncDataTree<IChatResponseProgressFileTreeData, IChatResponseProgressFileTreeData, void>> {
+		const object = this._pool.get();
+		let stale = false;
+		return {
+			object,
+			isStale: () => stale,
+			dispose: () => {
+				stale = true;
+				this._pool.release(object);
+			}
+		};
+	}
+}
+
+class ChatListTreeDelegate implements IListVirtualDelegate<IChatResponseProgressFileTreeData> {
+	static readonly ITEM_HEIGHT = 22;
+
+	getHeight(element: IChatResponseProgressFileTreeData): number {
+		return ChatListTreeDelegate.ITEM_HEIGHT;
+	}
+
+	getTemplateId(element: IChatResponseProgressFileTreeData): string {
+		return 'chatListTreeTemplate';
+	}
+}
+
+class ChatListTreeCompressionDelegate implements ITreeCompressionDelegate<IChatResponseProgressFileTreeData> {
+	isIncompressible(element: IChatResponseProgressFileTreeData): boolean {
+		return !element.children;
+	}
+}
+
+interface IChatListTreeRendererTemplate {
+	templateDisposables: DisposableStore;
+	label: IResourceLabel;
+}
+
+class ChatListTreeRenderer implements ICompressibleTreeRenderer<IChatResponseProgressFileTreeData, void, IChatListTreeRendererTemplate> {
+	templateId: string = 'chatListTreeTemplate';
+
+	constructor(private labels: ResourceLabels, private decorations: IFilesConfiguration['explorer']['decorations']) { }
+
+	renderCompressedElements(element: ITreeNode<ICompressedTreeNode<IChatResponseProgressFileTreeData>, void>, index: number, templateData: IChatListTreeRendererTemplate, height: number | undefined): void {
+		templateData.label.element.style.display = 'flex';
+		const label = element.element.elements.map((e) => e.label);
+		templateData.label.setResource({ resource: element.element.elements[0].uri, name: label }, {
+			title: element.element.elements[0].label,
+			fileKind: element.children ? FileKind.FOLDER : FileKind.FILE,
+			extraClasses: ['explorer-item'],
+			fileDecorations: this.decorations
+		});
+	}
+	renderTemplate(container: HTMLElement): IChatListTreeRendererTemplate {
+		const templateDisposables = new DisposableStore();
+		const label = templateDisposables.add(this.labels.create(container, { supportHighlights: true }));
+		return { templateDisposables, label };
+	}
+	renderElement(element: ITreeNode<IChatResponseProgressFileTreeData, void>, index: number, templateData: IChatListTreeRendererTemplate, height: number | undefined): void {
+		templateData.label.element.style.display = 'flex';
+		if (!element.children.length && element.element.type !== FileType.Directory) {
+			templateData.label.setFile(element.element.uri, {
+				fileKind: FileKind.FILE,
+				hidePath: true,
+				fileDecorations: this.decorations,
+			});
+		} else {
+			templateData.label.setResource({ resource: element.element.uri, name: element.element.label }, {
+				title: element.element.label,
+				fileKind: FileKind.FOLDER,
+				fileDecorations: this.decorations
+			});
+		}
+	}
+	disposeTemplate(templateData: IChatListTreeRendererTemplate): void {
+		templateData.templateDisposables.dispose();
+	}
+}
+
+class ChatListTreeDataSource implements IAsyncDataSource<IChatResponseProgressFileTreeData, IChatResponseProgressFileTreeData> {
+	hasChildren(element: IChatResponseProgressFileTreeData): boolean {
+		return !!element.children;
+	}
+
+	async getChildren(element: IChatResponseProgressFileTreeData): Promise<Iterable<IChatResponseProgressFileTreeData>> {
+		return element.children ?? [];
+	}
+}
