+++ src/vs/workbench/contrib/chat/browser/chatContentParts/media/chatConfirmationWidget.css
